pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kaisho"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "apscheduler>=3.10",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "watchfiles>=0.21",
    "python-multipart>=0.0.9",
    "youtube-transcript-api>=1.0",
]

[project.scripts]
kai = "kaisho.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["kaisho*"]
